Reduce RHmin in upper stratosphere and lower mesosphere to be below stratopause RH (#1186)
* Reduce minimum RH above 3 mbar by a factor of ten.

Old value led to q > 10 ppm near polar stratopause.
This version keeps specific humidity near its values in the rest of the stratosphere.

* DOC: Add changelog entry

They may want an upper bound to the change at some point, but I don't know that we get far enough into the mesosphere for that to be an issue

* BUG: Use floating-point literals to reduce stratopause RHmin

* add hera RT log

* add hercules RT log

* add orion RT log and cmp path fix gfs

* Update intel.yml update ubuntu version 22.04

---------

Co-authored-by: WenMeng-NOAA <48260754+WenMeng-NOAA@users.noreply.github.com>
Co-authored-by: FernandoAndrade-NOAA <fernando.andrade-maldonado@noaa.gov>
Co-authored-by: Fernando Andrade - NOAA <121964418+FernandoAndrade-NOAA@users.noreply.github.com>